pyproject.toml
[project]
name = "detect-face"
version = "0.1.0"
description = ""
authors = [
    {name = "Bittu Kumar",email = "bitweb24@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "deepface (>=0.0.93,<0.0.94)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "tf-keras (>=2.19.0,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "ultralytics (>=8.3.150,<9.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "keras (>=3.10.0,<4.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "webcolors (>=24.11.1,<25.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
